refactor: restructure some things
